pyproject.toml
[project]
name = "DeepImageSearch"
version = "3.0.0"
description = "AI-powered image search engine with multimodal embeddings, text-to-image search, LLM captioning, and agentic RAG integration."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nilesh Verma", email = "me@nileshverma.com" }
]
keywords = [
    "Deep Image Search Engine",
    "AI Image Search",
    "Image Search Python",
    "CLIP Image Search",
    "Multimodal Search",
    "Vision RAG",
    "Agentic Search",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "open-clip-torch>=2.26.0",
    "timm>=1.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
chroma = ["chromadb>=0.5.0"]
qdrant = ["qdrant-client>=1.9.0"]
llm = ["openai>=1.50.0"]
mcp = ["mcp[cli]>=1.0.0"]
langchain = ["langchain-core>=0.3.0"]
postgres = ["psycopg2-binary>=2.9.0"]
all = [
    "chromadb>=0.5.0",
    "qdrant-client>=1.9.0",
    "openai>=1.50.0",
    "mcp[cli]>=1.0.0",
    "langchain-core>=0.3.0",
    "psycopg2-binary>=2.9.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/TechyNilesh/DeepImageSearch"
Repository = "https://github.com/TechyNilesh/DeepImageSearch"
Issues = "https://github.com/TechyNilesh/DeepImageSearch/issues"

[project.scripts]
deep-image-search-mcp = "DeepImageSearch.agents.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["DeepImageSearch"]

[tool.ruff]
target-version = "py310"
line-length = 120
